requirements.txt
psycopg2-binary==2.7.4
PyYAML==3.12
autopep8==1.4.3